pyproject.toml
[project]
name = "resume-parse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=23.2.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "black>=25.11.0",
    "fastapi>=0.104.0",
    "langchain>=0.1.0",
    "langchain-openai>=1.0.3",
    "langgraph>=0.0.50",
    "loguru>=0.7.0",
    "openai>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymupdf>=1.23.0",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.28.0",
    "uvicorn>=0.24.0",
]
